Implemented Toggle to Keep Craft Commands after Dogfight
Currently when a patrolling/escorting/chasing craft is being engaged by a hunter-killer
the craft will return to base after the successfully crashing/destroying the hunter-killer.
This means that the original command has to be to the craft again.
This implementation adds a toggle, so the craft will not return to base unless
it has achieved its final mission.