"Prison Cleanup" button
There is a new button "Prison Cleanup" in the Alien Containment GUI.

The button appears only if the interrogated/transferred prisoners
are necessary to resolve the situation with insufficient prison space.
If the situation can be resolved without them, button will not appear.

Pressing the button cancels all ongoing interrogations (for the given
prison type) and all prisoner transfers (of the given prison type).
Incoming prisoners will be treated as if they arrived already,
so that the player can choose to get rid of them.