Several fixes
If both brutal and leeroyjenkins are set for a unit, brutal gets the priority.
Will no longer try to move units that have no energy.
Energy is now considered for determining how far we can walk.
Fixed an issue that caused melee-attacks not being considered when the facing wasn't right.