Behavioural rework
In Non-omniscient-mode visiting or seeing a tile that an enemy unit was suspected to be at will now make the AI realize that the enemy is not there and it will keep searching.

Lowered the prerequisite of pre-priming grenades. That will happen much more frequently now.

Completely new hide&peak default-behaviour that is much more about safety and getting the first-strike. AI will avoid open territory but still seek for spots to perform this behaviour on.
However, there is a morale-based "flipswitch" that turns the AI back into sweeping-mode, where it chase down enemy-units.

This sweeping-mode can be enforced by mods to always be active by enabling the unit's Leeroy-flag.

The AI switching between two modes of operation based on morale creates more dynamic and immersive gameplay than sticking to one behaviour would have.

Fixed that clearsight and quicklineoffire-functions did not consider terrain-elavation which could cause units falsely thinking they are hidden when standing on a hill, as they are common on desert and forrest maps.

Fixed a crash that could happen at the end of a mission and was introduced by one of the previous versions.