Update ProjectileFlyBState.cpp
Fixed an engine-issue that led to the AI sometimes doing nothing when a grenade-throw was ordered but it would hit a wall.
This bug also impacts the player in a way that the throw simply doesn't happen. However, the player could just click again and thus was guaranteed never to hit any walls.
This fix will now allow bad throws against walls to happen for both sides.