Several fixes
Fixed an issue that prevented AI-units that were above carrying-capacity from moving under certain circumstances.
Non-omniscient AI is now aware of where players units are to be expected at the beginning of a mission.
After the search-algorithm actually worked as intened, it was realized that it's pretty bad so it was changed to assume the enemy units were walking as far as they could instead of to the closest tile that is out of vision.