Update AIModule.cpp
Found a better compromise between AI-turn-processing-time and moving in an ideal order. Turn-times should be a lot lower and units that see an equal amount of enemies now move in the order of who can reach the most tiles. This way they can still correctly move the units near the exit of the Skyranger first.

AI-units no longer assume they can attack from their current position in movement-logic. If they made it to the movement-logic they obviously couldn't attack and their tile is a false-positive.