Further improvements
Fixed an issue that prevented the non-omniscient AI from noticing when a tile was checked for enemy presence and made it think the enemy is still there even when it saw the tile to be empty.

AI will no longer take a peak when it thinks or knows the enemy is too far away to get a clean shot. This shall only affect behaviour with enabled UFOExtender: Accuracy.

The mechanism of how the non-omniscient AI concludes a new search-target should now be way faster while achieving roughly the same results.