waking up from having passed
When a friend of a unit gets hit or the amount of visible opponents has changed AI-controlled units, that had already passed their turn wake up again to see if the situation has changed in a way that they can do more than when they were checking last. This should help in scenarios such as unboarding the Skyranger in turn 1 while immediately encountering opponents.