Update AIModule.cpp
AI now realizes the difference between not having been able to attack because there's no line of fire despite seeing the target and not being able to attack because there's a door blocking the way. This fixes an issue where the AI was walking back and forth on one side of a door instead of opening it.