Tweaks & fixes
Fixed an issue introduced in 3.5.0 where the AI overwrote when they would have a line of fire. This made them think they find tile to attack from and so they'd usually just run away unless they already had a line of fire where they already were are. This dramatically reduced their performance. If you've been using 3.5.0, please update immediately.

AI should now be better at guessing what is in the line of fire of an enemy unit.

AI-units who's primary-weapon is a melee-weapon but who also have grenades will now consider themselves a melee-unit, not a ranged-unit.

AI will now deliberately avoid the starting-location of the player when it hasn't seen any units of the player out in the open.

AI now more bold to advance towards enemy units that are not at the starting-location of the player.

Fixed an issue where the AI would sometimes flee despite being able to attack.

Instead of giving tiles with smoke on it a flat 25% score-bonus the AI now gives a 2% score-bonus per level of smoke.

Reverted a prior change that made the AI overly cautious about avoiding friendly-fire. Tests showed this often did more harm than good.