Update ProjectileFlyBState.cpp
When a unit of the AI hits an enemy all friends of that unit who have already passed their turn will reconsider whether there's something they should do.

Any shot fired will now give away the position of the attacking unit to the AI, not just when the shot hits.