disable blind-fire
Enemies are no longer afraid of primed smoke-grenades dropped to the ground.

For now disabled all sorts of blind-firing of the AI other than for blaster-like-weapons.
Explanation: The fix to the grenade-blind-firing not working depends on something that can be considered cheating. If it wouldn't, there's a high potential to exploit it by making the enemies waste their grenades for no effect. Same is already true for regular attacks. Basically: Knowing how the blind-fire-works makes it exploitable. So it seemed better not to be done at all. There may be a comeback of it in the future but I need an idea of how to do it without cheating and without exploit-potential at the same time.