Weapon-pickup-range
Changed the scoring for what weapons the AI should pick up from int to float. This increases the distance at which the AI is willing to look for a dropped weapon on the ground tremendously as the score will no longer become 0 due to rounding.

Note: In this context I noticed that weapon-pickup, as it is currently coded only works for basic-alien-weaponry and not for earthling-stuff or mod-stuff. In a future rewrite I plan to make it much more agile and work with anything.