Update Options.cpp
The "Omniscience for Brutal-AI"-option is no longer enabled by default. I think the AI is now finally good enough to pose a decent challenge without having to cheat. If you like it to be even harder, you still have the option to enable omniscience, of course.

The default-Mod-validation-level has been turned down from 2 to 0. I think it's more reasonable to expect modders turn it up than it is people who don't know about it but who want to try a specific mod, to figure out they have to turn it down. I've even heard of people using older versions of OXCE because they couldn't get the mod to run on a newer one.