hiding logic
When stuck on the outside and without any cover in range, a unit of the AI will go into a more aggressive stance.

AI shall be more willing to hide in a good hiding-spot instead of standing around and doing nothing if it can.

AI will no longer be convinced it shall be able to attack an enemy-unit who's position it has only taken a guess for.

The edge of the map is no longer counted as cover from the respective side as noone will try to attack from outside of the map anyways.

The AI's willingness to peak will now depend on whether it previously was in a good position it can return to after peaking and whether the enemy was recently spotted somewhere in an exposed position.

In combination these changes lead to more of a "lure the player into a trap"-playstyle.