speedup
When an AI unit determined that it should wait, it now passes what unit it wants to wait for to the overarching process. That process then jumps directly to that unit instead of cycling through all the other units. This saves a lot of time. In a test run with Alien Colony Level one, it brought down the wait-time between turns from 48 to 21 seconds.

Reenabled centering the camera on the next selected unit when waiting in debug-mode as the epilepsy-inducing jumping-around from cycling through all units no longer happens.