new options for performance-optimization
The AI-options-menu now provides a new option instead of the Strafers-option:
Performance optimisation for huge maps

Here you can trade the AI's ability to navigate and cartograph the map for better turn-processing-performance in order to make massive maps like 120x120x10 with more than 100 enemies playable.

Navigation refers to the amount of tiles each unit considers to walk to. This is always the closest tiles from where it is currently located. In addition to that, flying units will reduce the number of tiles they want to fly to to half the amount for navigation. Note that if the unit can reach more tiles than the limit in a single turn, it'll still consider all of them.
Cartography refers to the tiles the AI remembers having seen, which plays a role in their scouting.

An idea for the future would be to come up with an automatic scaling-algorithm that determines good values based on the size of the map and the amount of enemies.