Update AIModule.cpp
Fixed an issue causing the omniscient AI not always positioning for an attack when it could.

Removed special handling for doors in positioning-code as that should no longer be necessary and can cause unwanted issues.

Fixed an issue that could cause not turning towards the right target after moving.

Fixed an issue with non-omniscient-AI using the omniscient-algorithm to determine an encircle-tile.