dynamic performance-adjustments
The AI-performance-optimization for huge maps now dynamically adjusts based on map-size and opponent count. It is normalized for maps with a size of 60x60x4 and 30 opponents. For more than that the adjustments kick in and limit the pathfinding-and tile-visibility of the AI.
Note that with hundreds of enemies it will still take a long time but the time and that the path-finding-distance will never drop below what a unit could reach in a single turn. But the time the AI spends on each enemy will be roughly constant leading to a more linear increase in turn-time based on the amount of enemies rather than an exponential one.