Leeroy fix
Brutal-AI is now treating the isLeeroyJenkins-flag in a way that is much more similar to the originally intended behaviour.
The unit will always try to get into melee-range and not use their ranged weapons offensively even if it can.

In order to get the previous behavior back, there now is a new flag for units that can be used in mods called "isAggressive". This is now what forced units to always move forwards and attack but without restricting what weapons can be used.

Consequentially the option "Charging behaviour for Brutal AI" was changed to "Force aggressive enemy-behaviour" and is a Yes or No-option rather than a range of values.