Basesecape Hangars reworked
This branch introduces two new features to OXCE+, related to Hangars &
Crafts management.

First, we can define different hangar types to host crafts.

A new key ("hangarType") is added for both crafts & facilities
rulesets (YAML) so the value for the craft must match the value at the
hangar to be allocated inside.

If that key is not defined, a default value (-1) is assigned, so both
crafts and hangar facilities behave as in original OXCE.

Also, an "Advanced Option" to show hangar types in crafts' UEFOPEDIA
entries is added. For Facilitie' UEFOPEDIA entries the modder should
add the text describing the hangar type.

Added "hangar Type" un "Stats for Nerds" screen for both crafts and
facilities.

Second, now is allowed to allocate more than 1 craft in Hangars (as
long as crafts have the same hangarType key than Hangar facility). All
crafts are shown in Basescape view inside their hangars, and you can
right-click an specific craft to go to its "craft screen".

For hangars with more than 1 craft, a new key ("craftSlots") is added
to the facilities ruleset (YAML) in order to define so many 'position
 vectors' [x,y,z] as crafts are allowed:

    crafts: 2
    craftSlots: # mandatory if crafts > 1; optional otherwise
      - [11, -11, 0]  # Distance [x,y] to the
      - [-11, 11, 0]  # central point

z value is not used.

If that key is not defined (and crafts=1), a default slot ([0, 0, 0])
is assigned, so it behaves as in "original" OXCE.

We have adjusted  "damageFacility", "destroyfacility"&"removeCraft"
at Base.cpp to consider a vector of crafts instead a single one (so,
 more than one craft and their allocated items and soldiers are
 considered)