Update alienDeployments.rul
Fixed several issues with the item-sets of enemies in Terror from the deep:
There will no longer be an unarmed high-ranking-enemy on terror-missions. It will instead use the same item-sets as the rank below it.
The enemies that previously brought a disruptor-pulse-launcher to attack the X-Com-base, now got the memo that it's not flooded and bring a weapon that actually works under dry conditions.