micro-mods
Added two standard-micro-mods. One for Ufo: Defense and one for Terror from the deep.

Both remove the item that X-Com can use to mind-control enemies while leaving mind-control for enemies and psi-screening-facilities intact so you can still screen for soldiers that are more resilient against mind-control.

Other than that they enable "Aggressive retaliation", "Enhanced dogfigt behaviour" and "Brutal AI".

The one for Terror from the deep, in additon makes the early-game a little less punishing by reducing the score-threshold for losing by 200 points. So you now must be below -500 instead of -300 twice in a row in order to lose.

The rationale for that was that mind-control makes the game ridiculously easy, regardless of what the AI would do with their units but can't because you already control them.