skiping reselect
Certain events cause units controlled by Brutal-AI to weak up again when they had already finished their turn, just in case there's now something to do. This was quite taxing on the turn-time as the entire AI-routine was run again. On big maps with many enemies and large sight-radius this could mean a lot of unneccessary AI-cycles. There now is a sanity-check in place for that where a very minimalist pathfinding-check is done to see if reactivaction could make sense at all. Basically when the unit cannot move anymore anyways and also doesn't have enough TUs for any attack, it will not be woken up again and not run another AI-cycle.