cover
Massive update to the target- and attack-mode-selection-algorithm.
Instead of just comparing the options at the current position it will now also compare options involving walking towards the target and using weapons at a shorter range with higher accuracy or melee-weapons if available. For that it will subtract the amount of time-units it needs walking over and sees how it impacts the amount of attacks it can do, the additional possibilities, whether it can sneak up or is in vision for reaction-fire and the impact on accuracy.

Damage vs. Armor comparisons now take into account what damage-distribution-model (UFO vs. TFTD) is being used.

When deciding where to take cover after peaking, cover is now evaluated relative to the best cover in range rather than any amount of cover being considered sufficient.

Units in sweep-mode no longer care about triggering reaction-fire or being too close to their friends.