Update AIModule.cpp
Fixed a rare but possible crash in line-of-fire determining-algorithm.

Units with limited weapon-range due to extender-accuaracy and lack of an aimed-shot, as it is often casse in WH40k/Rosigma are no longer forced into sweep-mode. They can now use appropriate logic by the new range-targeting and decide to seek/stay in cover or walk forward to take a shot.