Update AIModule.cpp
Importance of avoiding line of sight to assumed position enemy could reach when walking towards us is now considered much higher when looking for a safe spot to advance to.

When avoiding to stand close to allied units, units that are behind a wall are no longer taken into consideration.

When evaluating the importance of cover from certain directions the distance towards enemies in this direction is now taken into consideration.

When evaluation the importance of cover form certain direction the cover towards the direction 45° to the left and righ to that direction are now factored in with half the amount of cover-score each.