Omniscient fixes
Units in sweep-mode will once again be allowed to hide if they don't have enough TUs to attack anymore.

Fixed an issue that could cause AI to skip the rest of their turn instead of taking cover.

Fixed that AI would think it just needs to turn towards their target despite not having a line of fire.

Omniscient AI no longer assumes that it will be able to attack everything it has a line of fire to. This could lead to them standing out of position when the player used smoke-grenades.

AI now will assume that it will be able to see a unit when it is at minimum range to it.

The peaking-logic now works the same for omniscient and non-cheating-AI.

Fixed an issue with omniscient AI not being omniscient when it comes to knowing whether an enemy was out of position.