many smaller changes
AI now can decide to move forwards even without being able to discover anything new, when some of their units are already engaged in combat or in the situation where it earlier would have gone into sweep-mode.

AI will no longer refrain from peaking based on the maximum range of their weapons based on their current TUs but based on what it would be if it had full TUs.

Instead of sweep-mode being a hard on/off-switch the AI becomes gradually more willing to move forward faster as it gains a morale-advantage. It will usually still try to find some cover.

Non-aggressive-forced auto-play-AI has a fixed value of how much it wants to spend on moving forward vs. saving to hide. It will also always consider being in combat as to not leave units behind.

AI will now take their energy into account when it considers peaking and rather not do it, if the energy is too low.

AI will now move in to attack instead of pre-priming a grenade if it can attack.

Fixed an issue where AI would not always turn towards a door it wants to open.

Melee-units now take potential reaction-fire into account too and might refrain from peaking. It should not prevent them from trying to move in for a kill though.

When trying to hide always prefer locations with cover over those without and where the enemy couldn't reach.

The regular cover-seeking score which is usually used to get into some cover after moving forwards and not finding anything to attack will now prefer to conserve TUs and no longer has redundant score-modifiers.

AI shall more actively try to avoid standing in the blast-radius of its own grenades or in fire.