Fix equipment list not being properly updated
When loading an equipment list for a craft,
if a warning box was triggered due to missing items
then the item list would not be correctly updated to show
which items are equipped by the crew.

This was noticeable only with the alternate inventory mgmt option.

The bug was that the warning box caused the list update to be suppressed.