Alternate loadout system #2
https://openxcom.org/forum/index.php/topic,10998.0.html

When using `oxceAlternateCraftEquipmentManagement`,
entering the inventory screen from a craft will now show
all available equipment in the base.
Changes made in the soldier's inventory are then used to update
the items loaded on the craft, so that the craft has
all of the soldiers' items
(as well as any unequipped items already loaded onto the craft).

Previously, entering the inventory screen from a craft would show
only the items that were already on-board the craft, and changes
made to soldiers inventory would not affect the items on the craft.