Update AIModule.cpp
AI no longer consideres something as cover when it doesn't fully block vision.
AI now is aware that when it's head peaks over some cover, for example when it stands on a slope, this cannot be considered as being in cover either.
When a unit controlled by the AI doesn't have any valid cover in range according to these new definitions and has a line of fire to an enemy or assumed position of an enemy it will now go into sweep-mode.