New moving-logic
The movement-logic of the AI has been largly reorganized.
The only thing that still works like before is how the AI positions itself to attack units it spotted in the current turn.
The second priority is now to move towards the closest full cover to the enemy, that the enemy cannot reach during its turn. For non-omniscient this is based on the assumption where the enemy is. An exception to that is when the unit already is at the closest full-cover position. It will then not relocate to a position further away.
Peaking has been vastly reduced in range to preserve more TUs for reaction fire. When a unit already is in the closest cover then moving one tile to peak and then back to the cover should be enough.
Fleeing will only be done when no adequate cover can be found.

The algorithm that makes assumptions about where an enemy could have moved to, when the tile it was assumed to be at is confirmed to not contain the enemy was updated to avoid vastly implausible results. It's now simply the tile closest to the previous location that wasn't explored in the current turn.