Upgrades
When enclosing while in cover, instead of being okay with any kind of cover, there's now three different levels of cover that will be prioritized if available:

The highest priority is a tile that provides cover to all directions enemies currently are known to be at and it should not be visible from the furthest tile the closest enemy could walk to.

If no such tile is found, the best tile with cover to all directions enemies are known to be will be used.

If no such tile is found, the tile with the best compromise of cover will be used.

Peaking-logic has been vastly improved and also was split into two different priorities.
The first priority is to peak towards a tile that the closest enemy is suspected to be at. The target tile is the closest tile which should have a line of sight towards the enemy.
If no such tile was found, peaking will be performed towards the first tile that is no longer in line of sight from our current position when pathing towards the position an enemy is suspected to be at.
The peaking tile needs to be reachable while keeping TUs and energy above 50% and the clostest tile that fulfills the criteria is chosen.

Fixed an issue that caused considering units without a line of sight as unit to look turn towards instead of the path that unit would have to take.