Update AIModule.cpp
Fixed an issue related to not looking in the right direction.

Targets assumed to be in smoke will no longer be mentally relocated upon seeing the tile.

When wanting to peak at a target assumed to be in smoke the unit will actually walk again to where it thinks the target is instead of just looking at its direction.