more upgrades
AI now determines a specific location to walk to and peak along the path to an enemy if it can't peak to the location the enemy is assumed to be.

AI will no longer head prioritize heading for cover when it already is in decent cover. Instead it should prioritize peaking in this situation. It will still prioritize taking cover if it is exposed and can do so.

AI will no longer flee with units when these units are not exposed.

AI shall now no longer consider doors as decent cover. They will still be considered low-quality-cover if nothing else is available.

Fixed another issue causing the AI not to look where it was supposed to look after moving.