Update AIModule.cpp
Fixed an issue where large units wouldn't know to open a door with an enemy on the other side in order to attack it, when the door was not adjecent to the tile their position is stored in.

AI will now first try to find great cover, then good cover and then okay cover.
AI will now always prefer to get closer to their target even for lower cover-tiers.
Fleeing was completely removed. When a unit doesn' find any at least okay cover within it's reach it will go forward with the only exception being large units who think the path to their target is too narrow.

Fixed an issue that could cause enemies to temporarily become passive in a x-com-base-defense-mission.