Update AIModule.cpp
Fixed an issue that caused seeking cover not being taken into coniseration when too close to the enemy.

Opening a door that an enemy is expected to hide behind now has the same leftover timeunit- and energy-requirements as other forms of peaking.

No longer consider hiding-spots that are further away from the current location than from the position a unit gravitates to as this could cause seeking cover from an enemy by walking past that enemy.