Update AIModule.cpp
When after peaking no tile can be reached that qualifies as "great cover", instead of going to position closest to the enemy that is considered "good cover", units controlled by the AI will now instead go to the closest position from their current position that is "okay cover".
The rationale is that it's much better to be in a slightly worse position and have enough TUs for reaction fire than being in a position that is almost as bad but being guaranteed not to be able to reaction-fire.

Tiles next to doors are no longer considered as viable option for cover-taking. This is because door-opening will trigger mutual surprise and the unit would likely be visible right after.