Update AIModule.cpp
Brought back a mechanism for the AI to "clean up" the map when player-units panic. However, it works differently now. Instead of just turning on sweep-mode for all units once a certain moral-discrepancy is reached, the AI will memorize which units have panicked and ignore these units when it comes to avoiding their line of fire.