endless-loop
Fixed a rare endless-loop related to AI-unit-move-order.
The value of how far the currently looked at unit can move wasn't updated in the reachable-positions-cache and could lead to an inconsistency when it was later compared to the cached value for the same unit from the other unit's position.
Now the cache gets forcefully updated whenever the unit considers taking it's turn which shall avoid these inconsistencies.