Close cover
In order to determine the best tile to go to when cover is desired but great cover cannot be reached there is now a new algorithm in place:

Similarly to the recognition of great cover there will be a tile-vision-check from all supposed enemy-reachable tiles to all potential tiles to go to. A score is determined where the starting value is the amount of supposed enemy-reachable tiles and where for each successfull tile-vision-check one is subtracted. The resulting target-tile is the one that had the highest remaining score, meaning it will be the one that will be hardest and least likely to discover.

The cuddleAvoidModifier is not applied to this kind of score.

Walking through someone's line of sight while seeking cover will no longer reduce the score of the target-tile as this often resulted in units just staying where they are in order to avoid reaction-fire just to get killed anyways for standing in a terribly exposed-position.

The results of the tile-visibility-checks are now cached. This cache is reset every time terrain gets destroyed or doors are opened.