New peak behavior
Instead of always prefering to try and peak in a way that should establish visual contact to the assumed location of the enemy, the AI now prefers peaking towards the peak-position, which is the closest tile toward the assumed location of the enemy it couldn't see from its position. Whether this reveals the enemy depends on how close the enemy has already gotten. This leaves more TUs for going back to a saver location or reaction-fire. But most importantly, it should prevent intentional or uninentional exploits of AI-units always trying to establish a line of sight to units and then being stranded in a bad location.