new options
Added a new option and changed 2 existing ones.
The aggression for enemies and autoplay now has 3 steps instead of two.
As of yet the balanced behavior is much closer to the defensive one. The reason for that is that during benchmark-testing it became apparent that more aggression usually lead to worse results.
The new option will use the aggression of the unit from the units.rul-file to determine the aggressiveness.