aimodule-changes
When deciding which enemy-unit to look for, the turn that unit was last spotted always took precedence. This allowed to distract the AI with a unit that is very far away.

Now it uses distance and turn at the same time for the scoring. Each turn that has passed adding the maximum amount of time-units a unit has to their perceived pathfinding-distance when it comes to how close it likely is.

Fixed an issue with going back and forth while hiding caused by forcing the AI to ignore the tile it currently stands on for its considerations.

The cover-scores now use some sort of heat-map around the supposed enemy-locations instead of the previous more binary approach. Meaning a tile is considered as safer depending on how many time-units an enemy would have after moving to a tile from where they can look at the AI's unit. The distance also flows into the formula for that.

Basically the AI will push towards the player's units as long as it is safe and will compromise a bit of the safety in order to get closer. However, the weights are still favoring better cover over being closer as this generally showed better results in testing.

The balanced approach will now once again prefer direct-peaking over indirect-peaking if possible.

Fixed an issue where an explosion with an efficacy of less than 1 would be considered as not worth causing. The explosion-efficacy now got it's own factor for the score and was separated from the number of shots, which was rounded down to the next full number.