more sophistication
If nearby allied units could also reach a tile their remaining time-units will be subtracted from the fear of being discovered on that tile. This makes AI-units more willing to venture towards enemies when they have a numerical advantage as fewer enemies also project less fear to be discovered.

The visibility-reduction of smoke is now also taken into account for the fear of visibility. It is divided by 1 + smoke-level / 3. This means the AI will now deliberately use your smoke against you.

The difference between AI-aggressiveness 1 and 2 is now that level 1 will always avoid standing in the line of fire of known or assumed enemies whereas level 2 will be okay with it if the projected fear of discovery is low enough.