Update AIModule.cpp
Removed impact of smoke again as it was then possible to exploit the AI by luring it to certain locations and then know exactly where to look for them on your next turn.

The least aggressive variant now once again tries to make itself undiscoverable and doesn't become more courageous in higher numbers.

So now the three levels of aggressiveness shall all be pretty distinct.