Smoke stuff
When calculating bravery to counteract the fear of enemies, the unit no longer gains bravery from itself, only from other units on their team.

The algorithm that checks whether an enemy should have been detected at a location where it was suspected to be but was not, now is aware of the viewing-distance inside of smoke. This means the location will be updated in this case instead of continuously staring at an empty location.

AI now has a slight fear of smoke and will also attempt to reduce line of sight towards smoke-tiles when hiding.

Removed a piece of coded I didn't remember what it was supposed to do and which under certain circumstances would overrule where the AI wanted to walk to.

Balanced Aggressiveness will now prefer to break line of sight with known units, if possible.

AI will now always want to spend up to half of it's TUs to peak into smoke when it would end up inside of the smoke itself while doing so.

When traceAI is enabled and you use Autoplay, the player-units will constantly produce smoke-clouds around them for testing-purposes. This is more of a benchmarking-feature and shouldn't really have much impact on actual gameplay.

When AI-performance-optimization is enabled flying units will just assume that they would be visible on tiles more than 2 levels above the closest floor instead of doing the calculations. This should save some processing-time on high maps with high-TU-flying-units with barely any negative impact.