fixes
Fixed an issue where the location of an enemy would be guessed despite the enemy having been spotted on the current turn.

Fixed an issue that caused units to shoot at enemies in close-quater-combat when standing diagonlly to them caused by using an int to determine distance and that int rounding up to 2, which was not less than 2, which was what was checked against.

Melee-dodge is now subtracted from the hit-chance for melee-attacks when the AI considers using melee against enemies.