blind-grenades
Fixed an issue that caused the AI to no longer be able to select it's units in the order it wanted.

The location of nearby allies no longer boosts the AI's bravery when it comes to choosing hiding spots. Turned out there simply is no benefit to compromising how well a unit is hidden, when it's the enemies' turn next.

Fixed an issue that caused the AI to preprime grenades in situations where it shouldn't. For example when standing right within the reach of the enemy.

The AI now determines an urgency to get rid of it's preprimed grenades. The more allies it is surrounded by, the more urgent does it want to get rid of the preprimed-grenade.

One of the things this urgency does is to boosts the score of throwing grenades during attack-option-selection.

Fixed an issue that prevented the AI to use non-explosive weapons against targets that are standing in the blast-radius of a grenade.

Blind-grenading is back and it is no longer cheating. The conditions under which the AI considers blindly throwing a grenade are as follows:
The urgency to get rid of a pre-primed grenade must be increased by having at least one ally standing in the blast-radius of the unit holding it.
There must be an available throwing-arc to a position that an enemy unit has been spotted within the previous turn. Note that all sorts of attacking does get units spotted to the AI.
No allied units must be within the blast-radius.
If there's several possibilities for that, the one that is furthest away gets chosen in order to minimize the chance of friendly-fire.
If you suspect the AI to be able to blind-grenade make sure not to stay in the location you attacked from.