Item drops
Items appearing or exploding on the ground are now assumed to hint the position of an enemy to the AI, similarly to opening doors and throwing or shooting.

So for example dropping primed smokes from your inventory to your feet will no longer be ignored by the AI when it comes to concluding where your units might be.