peaking
Smoke-peaking will now also be performed when the closest enemy is assumed to be in smoke.

When the AI thinks it should be able to get visual contact to an enemy unit, it will extend the peaking-distance to up to half it's TUs.

It will still be conservative with it's peaking-TUs, when it thinks it likely can't spot the enemy anyways.

AI should avoid hiding next to regular doors, that it would then often close and open over and over.