Update AIModule.cpp
AI now shall move units that are standing in a doorway first.

This is necessary because units halt their movement after opening a door. After that they can have a low priority to move again. Units that are blocked by the unit standing in the doorway might now decide there is no path towards the enemy and just waste their TUs to walk to another tile in the craft. If the unit in the doorway immediately moves again, the path can be cleared and all other units can also leave the craft.